update spec (+node config etc) separate test files update http server to allow get-tx functionality
