+++ services/publicapi/basic_test.go
@@ -29,7 +29,7 @@ func TestPublicApiSendTx_TranslateTransactionStatusToHttpCode(t *testing.T) {
 		{"TRANSACTION_STATUS_REJECTED_GLOBAL_PRE_ORDER", protocol.REQUEST_STATUS_REJECTED, protocol.TRANSACTION_STATUS_REJECTED_GLOBAL_PRE_ORDER},
 		{"TRANSACTION_STATUS_REJECTED_VIRTUAL_CHAIN_PRE_ORDER", protocol.REQUEST_STATUS_REJECTED, protocol.TRANSACTION_STATUS_REJECTED_VIRTUAL_CHAIN_PRE_ORDER},
 		{"TRANSACTION_STATUS_REJECTED_SMART_CONTRACT_PRE_ORDER", protocol.REQUEST_STATUS_REJECTED, protocol.TRANSACTION_STATUS_REJECTED_SMART_CONTRACT_PRE_ORDER},
-		{"TRANSACTION_STATUS_REJECTED_TIMESTAMP_PRECEDES_NODE_TIME", protocol.REQUEST_STATUS_REJECTED, protocol.TRANSACTION_STATUS_REJECTED_TIMESTAMP_PRECEDES_NODE_TIME},
+		{"TRANSACTION_STATUS_REJECTED_TIMESTAMP_PRECEDES_NODE_TIME", protocol.REQUEST_STATUS_REJECTED, protocol.TRANSACTION_STATUS_REJECTED_TIMESTAMP_AHEAD_OF_NODE_TIME},
 		{"TRANSACTION_STATUS_REJECTED_CONGESTION", protocol.REQUEST_STATUS_CONGESTION, protocol.TRANSACTION_STATUS_REJECTED_CONGESTION},
 	}
 	for i := range tests {
