--- services/publicapi/test/public_api_test.go
@@ -1,259 +1,0 @@
-package test
-
-import (
-	"context"
-	"github.com/orbs-network/go-mock"
-	"github.com/orbs-network/orbs-network-go/config"
-	"github.com/orbs-network/orbs-network-go/crypto/digest"
-	"github.com/orbs-network/orbs-network-go/instrumentation/log"
-	"github.com/orbs-network/orbs-network-go/services/publicapi"
-	"github.com/orbs-network/orbs-network-go/test"
-	"github.com/orbs-network/orbs-network-go/test/builders"
-	"github.com/orbs-network/orbs-spec/types/go/protocol"
-	"github.com/orbs-network/orbs-spec/types/go/protocol/client"
-	"github.com/orbs-network/orbs-spec/types/go/services"
-	"github.com/orbs-network/orbs-spec/types/go/services/handlers"
-	"github.com/stretchr/testify/require"
-	"os"
-	"testing"
-	"time"
-)
-
-func TestSendTransaction_CallsTxPool(t *testing.T) {
-	test.WithContext(func(ctx context.Context) {
-		harness := newPublicApiHarness(ctx, 1*time.Millisecond)
-
-		harness.txpMock.When("AddNewTransaction", mock.Any).Return(&services.AddNewTransactionOutput{}).Times(1)
-
-		harness.papi.SendTransaction(&services.SendTransactionInput{
-			ClientRequest: (&client.SendTransactionRequestBuilder{
-				SignedTransaction: builders.Transaction().Builder()}).Build(),
-		})
-
-		ok, err := harness.txpMock.Verify()
-		require.True(t, ok, "should have called the txp func")
-		require.NoError(t, err, "error happened when it should not")
-	})
-}
-
-func TestSendTransaction_AlreadyCommitted(t *testing.T) {
-	test.WithContext(func(ctx context.Context) {
-		harness := newPublicApiHarness(ctx, 1*time.Millisecond)
-
-		harness.txpMock.When("AddNewTransaction", mock.Any).Return(&services.AddNewTransactionOutput{
-			TransactionStatus:  protocol.TRANSACTION_STATUS_DUPLICATE_TRANSACTION_ALREADY_COMMITTED,
-			TransactionReceipt: builders.TransactionReceipt().Build(),
-		}).Times(1)
-
-		result, err := harness.papi.SendTransaction(&services.SendTransactionInput{
-			ClientRequest: (&client.SendTransactionRequestBuilder{
-				SignedTransaction: builders.Transaction().Builder()}).Build(),
-		})
-
-		require.NoError(t, err, "error happened when it should not")
-		require.NotNil(t, result, "Send transaction returned nil instead of object")
-	})
-}
-
-func TestSendTransaction_BlocksUntilTransactionCompletes(t *testing.T) {
-	test.WithContext(func(ctx context.Context) {
-		harness := newPublicApiHarness(ctx, 1*time.Second)
-
-		txb := builders.Transaction().Builder()
-		harness.OnAddNewTransaction(func () {
-			harness.papi.HandleTransactionResults(&handlers.HandleTransactionResultsInput{
-				TransactionReceipts: []*protocol.TransactionReceipt{builders.TransactionReceipt().WithTransaction(txb.Build().Transaction()).Build()},
-			})
-		})
-
-		result, err := harness.papi.SendTransaction(&services.SendTransactionInput{
-			ClientRequest: (&client.SendTransactionRequestBuilder{
-				SignedTransaction: txb,
-			}).Build(),
-		})
-
-		require.NoError(t, err, "error happened when it should not")
-		require.NotNil(t, result, "Send transaction returned nil instead of object")
-	})
-}
-
-func TestSendTransaction_BlocksUntilTransactionErrors(t *testing.T) {
-	test.WithContext(func(ctx context.Context) {
-		harness := newPublicApiHarness(ctx, 1*time.Second)
-
-		txb := builders.Transaction().Builder()
-		txHash := digest.CalcTxHash(txb.Build().Transaction())
-
-		harness.OnAddNewTransaction(func () {
-			harness.papi.HandleTransactionError(&handlers.HandleTransactionErrorInput{
-				Txhash:            txHash,
-				TransactionStatus: protocol.TRANSACTION_STATUS_REJECTED_TIMESTAMP_WINDOW_EXCEEDED,
-			})
-		})
-
-		result, err := harness.papi.SendTransaction(&services.SendTransactionInput{
-			ClientRequest: (&client.SendTransactionRequestBuilder{
-				SignedTransaction: txb,
-			}).Build(),
-		})
-
-		require.NoError(t, err, "error happened when it should not")
-		require.NotNil(t, result, "Send transaction returned nil instead of object")
-		require.Equal(t, protocol.TRANSACTION_STATUS_REJECTED_TIMESTAMP_WINDOW_EXCEEDED, result.ClientResponse.TransactionStatus(), "got wrong status")
-	})
-}
-
-func TestGetTransactionStatus_CallsTxPool(t *testing.T) {
-	test.WithContext(func(ctx context.Context) {
-		harness := newPublicApiHarness(ctx, 1*time.Second)
-		harness.OnGetTransactionSetPending()
-
-		harness.papi.GetTransactionStatus(&services.GetTransactionStatusInput{
-			ClientRequest: (&client.GetTransactionStatusRequestBuilder{}).Build(),
-		})
-
-		ok, err := harness.txpMock.Verify()
-		require.True(t, ok, "should have called the txp func")
-		require.NoError(t, err, "error happened when it should not")
-	})
-}
-
-
-func TestGetTransactionStatus_GetCommitStatusFromTxPool(t *testing.T) {
-	test.WithContext(func(ctx context.Context) {
-		harness := newPublicApiHarness(ctx, 1*time.Second)
-
-		txb := builders.Transaction().Builder()
-		txHash := digest.CalcTxHash(txb.Build().Transaction())
-
-		harness.OnGetTransactionSetCommit()
-		result, err := harness.papi.GetTransactionStatus(&services.GetTransactionStatusInput{
-			ClientRequest: (&client.GetTransactionStatusRequestBuilder{
-				Txhash: txHash,
-			}).Build(),
-		})
-
-		require.NoError(t, err, "error happened when it should not")
-		require.NotNil(t, result, "get transaction status returned nil instead of object")
-		require.Equal(t, protocol.TRANSACTION_STATUS_COMMITTED, result.ClientResponse.TransactionStatus(), "got wrong status")
-		require.NotNil(t, result.ClientResponse.TransactionReceipt(), "got empty receipt")
-	})
-}
-
-func TestGetTransactionStatus_GetPendingStatusFromTxPool(t *testing.T) {
-	test.WithContext(func(ctx context.Context) {
-		harness := newPublicApiHarness(ctx, 1*time.Second)
-
-		txb := builders.Transaction().Builder()
-		txHash := digest.CalcTxHash(txb.Build().Transaction())
-
-		harness.OnGetTransactionSetPending()
-		result, err := harness.papi.GetTransactionStatus(&services.GetTransactionStatusInput{
-			ClientRequest: (&client.GetTransactionStatusRequestBuilder{
-				Txhash: txHash,
-			}).Build(),
-		})
-
-		require.NoError(t, err, "error happened when it should not")
-		require.NotNil(t, result, "get transaction status returned nil instead of object")
-		require.Equal(t, protocol.TRANSACTION_STATUS_PENDING, result.ClientResponse.TransactionStatus(), "got wrong status")
-		test.RequireCmpEqual(t, (*protocol.TransactionReceiptBuilder)(nil).Build(), result.ClientResponse.TransactionReceipt(), "Transaction receipt is not equal")
-	})
-}
-
-func TestGetTransactionStatus_GetTxFromBlockStorage(t *testing.T) {
-	test.WithContext(func(ctx context.Context) {
-		harness := newPublicApiHarness(ctx, 1*time.Second)
-
-		txb := builders.Transaction().Builder()
-		txHash := digest.CalcTxHash(txb.Build().Transaction())
-
-		harness.OnGetTransactionSetNotFound()
-		harness.OnGetTransactionSetBlockStorageFind()
-		result, err := harness.papi.GetTransactionStatus(&services.GetTransactionStatusInput{
-			ClientRequest: (&client.GetTransactionStatusRequestBuilder{
-				Txhash: txHash,
-			}).Build(),
-		})
-
-		require.NoError(t, err, "error happened when it should not")
-		require.NotNil(t, result, "get transaction status returned nil instead of object")
-		require.Equal(t, protocol.TRANSACTION_STATUS_COMMITTED, result.ClientResponse.TransactionStatus(), "got wrong status")
-		require.NotNil(t, result.ClientResponse.TransactionReceipt(), "got empty receipt")
-	})
-}
-
-type harness struct {
-	papi    services.PublicApi
-	txpMock *services.MockTransactionPool
-	bksMock *services.MockBlockStorage
-}
-
-func newPublicApiHarness(ctx context.Context, txTimeout time.Duration) *harness {
-	logger := log.GetLogger().WithOutput(log.NewOutput(os.Stdout).WithFormatter(log.NewHumanReadableFormatter()))
-	cfg := newPublicApiConfig(txTimeout)
-	txpMock := makeTxMock()
-	vmMock := &services.MockVirtualMachine{}
-	bksMock := &services.MockBlockStorage{}
-	papi := publicapi.NewPublicApi(ctx, cfg, txpMock, vmMock, bksMock, logger)
-	return &harness{
-		papi:    papi,
-		txpMock: txpMock,
-		bksMock: bksMock,
-	}
-}
-
-func newPublicApiConfig(txTimeout time.Duration) publicapi.Config {
-	cfg := config.EmptyConfig()
-	cfg.SetDuration(config.PUBLIC_API_SEND_TRANSACTION_TIMEOUT, txTimeout)
-	cfg.SetUint32(config.VIRTUAL_CHAIN_ID, uint32(builders.DEFAULT_TEST_VIRTUAL_CHAIN_ID))
-
-	return cfg
-}
-
-func makeTxMock() *services.MockTransactionPool {
-	txpMock := &services.MockTransactionPool{}
-	txpMock.When("RegisterTransactionResultsHandler", mock.Any).Return(nil)
-	return txpMock
-}
-
-func (h *harness)OnAddNewTransaction(f func ()) {
-	h.txpMock.When("AddNewTransaction", mock.Any).Times(1).
-		Call(func(input *services.AddNewTransactionInput) (*services.AddNewTransactionOutput, error) {
-			go func() {
-				time.Sleep(1 * time.Millisecond)
-				f()
-			}()
-			return &services.AddNewTransactionOutput{TransactionStatus: protocol.TRANSACTION_STATUS_PENDING}, nil
-		})
-}
-
-func (h *harness)OnGetTransactionSetPending() {
-	h.txpMock.When("GetCommittedTransactionReceipt", mock.Any).Return(&services.GetCommittedTransactionReceiptOutput{
-		TransactionStatus:  protocol.TRANSACTION_STATUS_PENDING,
-	}).Times(1)
-}
-
-func (h *harness)OnGetTransactionSetNotFound() {
-	h.txpMock.When("GetCommittedTransactionReceipt", mock.Any).Return(&services.GetCommittedTransactionReceiptOutput{
-		TransactionStatus:  protocol.TRANSACTION_STATUS_NO_RECORD_FOUND,
-	}).Times(1)
-}
-
-func (h *harness)OnGetTransactionSetCommit() {
-	h.txpMock.When("GetCommittedTransactionReceipt", mock.Any).Return(&services.GetCommittedTransactionReceiptOutput{
-		TransactionStatus:  protocol.TRANSACTION_STATUS_COMMITTED,
-		TransactionReceipt: builders.TransactionReceipt().Build(),
-	}).Times(1)
-}
-
-func (h *harness) OnGetTransactionSetBlockStorageFind() {
-	h.bksMock.When("GetTransactionReceipt", mock.Any).Return(
-		&services.GetTransactionReceiptOutput{
-			TransactionReceipt: builders.TransactionReceipt().Build(),
-			BlockHeight:        0,
-			BlockTimestamp:     0,
-		}).Times(1)
-}
-
-
